% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{air_data}
\alias{air_data}
\title{Air Pollution Data Measured in Disability-Adjusted Life Years (DALYs) (1990-2019)}
\format{
\subsection{\code{air_data}}{

A data frame with 720 rows and 5 columns:
\describe{
\item{country}{Lists the 23 countries in North America plus North America (totaling everything)}
\item{year}{Years of observation from 1990-2019}
\item{air_pollution}{total DALYs resulted from total air pollution: indoor and outdoor}
\item{household_pollution}{total DALYs resulted from indoor air pollution}
\item{outdoor_pollution}{total DALYs resulted from outdoor air pollution}
...
}
}
}
\source{
\url{https://ourworldindata.org/air-pollution}
}
\usage{
air_data
}
\description{
A subset of Air Pollution data filtered by countries in North America. DALYs is a metric used to assess disease burden. Data originally aggregated and maintained by Hannah Ritchie and Max Roser at \href{https://ourworldindata.org/air-pollution}{Our World in Data}.
}
\keyword{datasets}
