% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaced_functions.R
\name{displaced_areachart}
\alias{displaced_areachart}
\title{Internally Displaced Individuals Area Chart (2008-2021)}
\usage{
displaced_areachart(nation = "Trinidad and Tobago")
}
\arguments{
\item{nation}{Name of country (default: Trinidad and Tobago)}
}
\value{
An area chart demonstrating displaced individuals from 2008 to 2021 for a given country compared to the U.S
}
\description{
This function produces an area chart visualizing the amount of displaced individuals of a given country compared to the U.S (default: Trinidad and Tobago)
}
\examples{
displaced_areachart ()
displaced_areachart("Nicaragua")
}
