% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{displaced_by_disaster}
\alias{displaced_by_disaster}
\title{Internally Displaced Count in North America Based on Extreme Events Caused by Climate Change (2008-2021)}
\format{
\subsection{\code{displaced_by_disaster}}{

A data frame with 294 rows and 3 columns:

\describe{
\item{year}{Years of observation, from 2008-2021}
\item{country}{Lists the 23 countries in North America plus North America (totaling everything)}
\item{displaced_count}{Internal displaced people based on weather-related disasters such as floods or storms}
...
}
}
}
\source{
\url{https://data.worldbank.org/indicator/VC.IDP.NWDS}
}
\usage{
displaced_by_disaster
}
\description{
A subset of data on internal displacement associated with slow-onset environmental change and natural disasters filtered by countries in North America. This data comes from \href{https://www.internal-displacement.org/research-areas/Displacement-disasters-and-climate-change}{Internal Displacement Monitoring Centre (IDMC)} and organized by \href{https://data.worldbank.org/indicator/VC.IDP.NWDS}{The World Bank}.
}
\keyword{datasets}
