% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaced_functions.R
\name{displaced_linegraph}
\alias{displaced_linegraph}
\title{Internally Displaced Individuals Line Graph (2008-2021)}
\usage{
displaced_linegraph(nation = "United States")
}
\arguments{
\item{nation}{Name of country (default: United States)}
}
\value{
A line graph and area chart demonstrating the amount of people displaced from 2008 to 2021
}
\description{
This function produces a line graph depicting the toal amount of displaced individuals in a given country (default: USA)
}
\examples{
displaced_linegraph ()
displaced_linegraph("Haiti")
}
