% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaced_functions.R
\name{displacement_histogram}
\alias{displacement_histogram}
\title{Distribution of the range of people displaced, in a specific country, by natural disasters and climate change}
\usage{
displacement_histogram(nation = "United States")
}
\arguments{
\item{nation}{Name of country (default: United States)}
}
\value{
A histogram demonstrating the amount of people displaced based on the country selected
}
\description{
This function visualizes the distribution of internally displaced people in a specific country (default: USA)
}
\examples{
displacement_histogram()
displacement_histogram("Haiti")
}
