% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.summary.mpm.R
\name{export.summary.mpm}
\alias{export.summary.mpm}
\title{Export the summary output for an mpm object to a text file
Output the mpm summary to a tab-demimited file for processing by other
programs (Excel, Spotfire...)  If the filename is empty, return the data
instead of writing to file (useful for web services).}
\usage{
\method{export}{summary.mpm}(x, filename = "", ...)
}
\arguments{
\item{x}{object of class \code{summary.mpm} as produced by the function of
the same name}

\item{filename}{prefix used to name the output file following <filename>_xyz.txt}

\item{...}{further arguments; currently none are used}
}
\value{
the output is returned invisibly
}
\description{
Polar (spherical) coordinates are added if the \code{summary.spm} object
contains 2 (3) dimensions.
}
\seealso{
\code{\link{summary.mpm}}
}
\author{
Rudi Verbeeck, Tobias Verbeke
}
\keyword{manip}
