% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Famin81A.R
\docType{data}
\name{Famin81A}
\alias{Famin81A}
\title{Famin81A Data
Data with demographic indicators by region of the world}
\format{
A data frame with 18 observations on the following 5 variables.
\describe{ 
	\item{Region}{a factor with 18 levels giving the
  region} 
	\item{Population.growth}{a numeric vector}
 \item{Infant.Mortality}{a numeric vector; Infant mortality}
 \item{Dietary.Energy}{a numeric vector; Dietary Energy}
 \item{Productivity}{a numeric vector; Productivity} 
}
}
\description{
Famin81A Data
Data with demographic indicators by region of the world
}
\references{
Friday, L. and Laskey, R. (1989). The Fragile Environment, The
  Darwin College Lecture. Cambridge University Press, UK.
}
\keyword{datasets}
