% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_aas.R
\name{gamma_aas}
\alias{gamma_aas}
\title{Compute stage transition probabilities using the
“asymptotic age-within-stage structure” (AAS) method of
Kendall et al. (2019)}
\usage{
gamma_aas(s, l, t)
}
\arguments{
\item{s}{Numeric scalar. Survival rate for the focal stage.}

\item{l}{Numeric scalar. Current estimate of the population growth rate
(lambda).}

\item{t}{Numeric scalar. Duration of the focal stage (years).}
}
\value{
Numeric scalar. The fraction of survivors that transition to the
next stage.
}
\description{
Internal helper. Computes \code{gamma} (the fraction of survivors that
transition out of a stage) using the 'AAS' method of Kendall et al. (2019).
}
\references{
Kendall, B.E., Fujiwara, M., Diaz-Lopez, J., Schneider, S., Voigt, J. and Wiesner, S., 2019.
Persistent problems in the construction of matrix population models.
\emph{Ecological modelling}, 406, pp.33-43.
}
\keyword{internal}
