% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{gdp_plot}
\alias{gdp_plot}
\title{GDP(in current US$) for a specific country over the years}
\usage{
gdp_plot(nation = "United States", min_year = 1940, max_year = 2020)
}
\arguments{
\item{nation}{Name of country (default: United States)}

\item{min_year}{Start of time period (default: 1940 - also minimum year in dataset)}

\item{max_year}{End of time period (default: 2020 - also maximum year in dataset)}
}
\value{
A ggplot looking at GDP vs years for nation specified between time period specified.
}
\description{
This function plots the trend in GDP(in current US$) for a specific country (default: USA) over a period of time (default: 1940-2020)
}
\examples{
gdp_plot()
gdp_plot("Bahamas", min_year = 1980)
}
