% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_functions.R
\name{ghg_plot}
\alias{ghg_plot}
\title{Total Greenhouse Gas Emissions(in million tonnes) for a specific country over the years}
\usage{
ghg_plot(nation = "United States", min_year = 1940, max_year = 2020)
}
\arguments{
\item{nation}{Name of country (default: United States)}

\item{min_year}{Start of time period (default: 1940 - also minimum year in dataset)}

\item{max_year}{End of time period (default: 2020 - also maximum year in dataset)}
}
\value{
A ggplot looking at GHG vs years for nation specified between time period specified.
}
\description{
This function plots the trend in total greenhouse gas(GHG) emissions(in million tonnes) for a specific country (default: USA) over a period of time (default: 1940-2020)
}
\examples{
ghg_plot()
ghg_plot("Canada")
}
