% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_functions.R
\name{ghg_predict}
\alias{ghg_predict}
\title{Predict the Total Greenhouse Gas Emissions in million tonnes for a specific country for a specific year}
\usage{
ghg_predict(pred_yr, nation = "United States")
}
\arguments{
\item{pred_yr}{Year for predicted value}

\item{nation}{Name of country (default: United States)}
}
\value{
A printed message stating the predicted value for the year and nation specified
}
\description{
Predict the Total Greenhouse Gas Emissions in million tonnes for a specific country for a specific year
}
\examples{
ghg_predict(2033)
ghg_predict(2024, "Belize")
}
