\name{Golub}
\alias{Golub}
\alias{Golub}
\title{Golub (1999) Data...}
\description{Golub (1999) Data
Golub et al. (1999) data on gene expression profiles of 38 patients
suffering from acute leukemia and a validation sample of 34 patients.}
\details{The original data of Golub et al. (1999) were preprocessed as follows: genes
that were called 'absent' in all samples were removed from the data sets,
since these measurements are considered unreliable by the manufacturer of
the technology.  Negative measurements in the data were set to 1.

The resulting data frame contains 5327 genes of the 6817 originally reported
by Golub et al. (1999).}
\alias{Golub}
\alias{Golub.grp}
\docType{data}
\note{Luc Wouters et al. (2003), p. 1134 contains a typo concerning the
sample sizes of AML- and ALL-type and erroneously reported}
\references{Luc Wouters et al. (2003). Graphical Exploration of Gene
Expression Data: A Comparative Study of Three Multivariate Methods,
Biometrics, 59, 1131-1139.}
\keyword{datasets}

