% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map}
\alias{map}
\title{Generate a projected map of tracks shaded by behavioural state}
\usage{
map(m, proj = "lambert", params = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{m}{a fitted object of class mpm}

\item{proj}{a character string defining the map projection, default = "lambert"}

\item{params}{parameters specific to the supplied projection}

\item{xlim}{longitude range to plot}

\item{ylim}{latitude range to plot}
}
