% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tracks.r
\name{map_tracks}
\alias{map_tracks}
\title{Generate a projected map of tracks shaded by behavioural state}
\usage{
map_tracks(m, quickmap = FALSE, proj = "mercator", params = NULL,
  orientation = NULL, xlim = NULL, ylim = NULL, theme = "dark",
  landcol = grey(0.95))
}
\arguments{
\item{m}{a fitted object of class mpm}

\item{quickmap}{should a quick map be render using ggplot2::coord_quickmap, if so then proj, params and orientation arguments will be ignored}

\item{proj}{a character string defining the map projection, default = "lambert"}

\item{params}{parameters specific to the supplied projection}

\item{xlim}{longitude range to plot}

\item{ylim}{latitude range to plot}
}
