% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm.r
\name{mpm}
\alias{mpm}
\title{Random Walk with autocorrelation Filter}
\usage{
mpm(data, optim = c("nlminb", "optim"), verbose = FALSE)
}
\arguments{
\item{data}{a data frame of observations (see details)}

\item{optim}{numerical optimizer}

\item{verbose}{report progress during minimization}
}
\value{
a list with components
\item{\code{fitted}}{a dataframe of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{input dataframe}
\item{\code{tmb}}{the tmb object}
\item{\code{opt}}{the object returned by the optimizer}
}
\description{
Move Persistence Model
}
\details{
Fit a random walk with time-varying move persistence to location data
without measurement error

The input track is given as a dataframe where each row is a location a constant
time interval from the previous row, and columns:
\describe{
\item{'id'}{individual identification,}
\item{'date'}{observation time (POSIXct,GMT),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude.}
}
}
