% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vari_dur.R
\name{negent}
\alias{negent}
\title{Negative entropy (objective function for max-entropy optimization)}
\usage{
negent(p)
}
\arguments{
\item{p}{Numeric vector of probabilities summing to 1.}
}
\value{
Numeric scalar (negative Shannon entropy).
}
\description{
Negative entropy (objective function for max-entropy optimization)
}
\keyword{internal}
