% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_functions.R
\name{population_plot}
\alias{population_plot}
\title{Population for a specific country over the years}
\usage{
population_plot(nation = "United States", min_year = 1940, max_year = 2020)
}
\arguments{
\item{nation}{Name of country (default: United States)}

\item{min_year}{Start of time period (default: 1940 - also minimum year in dataset)}

\item{max_year}{End of time period (default: 2020 - also maximum year in dataset)}
}
\value{
A ggplot looking at population vs years for nation specified between time period specified.
}
\description{
This function plots the trend in population for a specific country (default: USA) over a period of time (default: 1940-2020)
}
\examples{
population_plot()
population_plot("El Salvador", max_year = 2010)
}
