% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mpm.R
\name{print.mpm}
\alias{print.mpm}
\title{Print Method for mpm Objects}
\usage{
\method{print}{mpm}(x, digits = 3, ...)
}
\arguments{
\item{x}{object of class mpm}

\item{digits}{minimum number of significant digits to be printed}

\item{...}{further arguments for the print method (for printing the contributions)}
}
\value{
x is returned invisibly
}
\description{
Print Method for mpm Objects
}
\seealso{
\code{\link{print.default}}
}
