% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{regional_grouping}
\alias{regional_grouping}
\title{North American countries grouped into regions and distinguished by economic class (2008-2021)}
\format{
\subsection{\code{regional_grouping}}{

A data frame with 24 rows and 3 columns:

\describe{
\item{year}{Years of observation, from 2008-2021}
\item{country}{Lists the 23 countries in North America plus North America (totaling everything)}
\item{region}{Regional grouping of countries in North America}
...
}
}
}
\source{
\url{https://data.worldbank.org/indicator/VC.IDP.NWDS}
}
\usage{
regional_grouping
}
\description{
A subset of data listing the regions and income-level classification (ex: high income). This data comes from \href{https://www.internal-displacement.org/research-areas/Displacement-disasters-and-climate-change}{Internal Displacement Monitoring Centre (IDMC)} and organized by \href{https://data.worldbank.org/indicator/VC.IDP.NWDS}{The World Bank}.
}
\keyword{datasets}
