% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_functions.R
\name{all_demographics}
\alias{all_demographics}
\title{Population, GDP, Total Greehouse Gas Emissions(GHG) over the years for a specific country}
\usage{
all_demographics(nation = "United States", ...)
}
\arguments{
\item{nation}{Name of country (default: United States)}

\item{...}{Other variables include min_year and max_year}
}
\value{
3 ggplots showing the relationships described above
}
\description{
This function plots 3 plots respectively showing (1) Population vs years (2) GDP(in current US$) vs years (3) Total GHG(million tonnes) vs years for a spcific country
}
\examples{
all_demographics()
all_demographics("Mexico", min_year = 2000)
}
