% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalys_function.R
\name{dalys_predict}
\alias{dalys_predict}
\title{Predict DALYs for a certain year for a certain country based on Total Greenhouse Gas Emissions}
\usage{
dalys_predict(pred_year, nation = "United States")
}
\arguments{
\item{nation}{}
}
\description{
Predict DALYs for a certain year for a certain country based on Total Greenhouse Gas Emissions
}
