% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalys_function.R
\name{dalys_predict}
\alias{dalys_predict}
\title{Predict Disability Adjusted Life Years and Total Greenhouse Gas emissions in million tonnes for a specific country in a future year}
\usage{
dalys_predict(pred_year, nation = "United States")
}
\arguments{
\item{nation}{}
}
\description{
Predict Disability Adjusted Life Years and Total Greenhouse Gas emissions in million tonnes for a specific country in a future year
}
