% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaced_functions.R
\name{displaced_areachart}
\alias{displaced_areachart}
\title{Internally Displaced Individuals Area Chart (2008-2021)}
\usage{
displaced_areachart(nation = "United States")
}
\arguments{
\item{nation}{Name of country (default: United States)}
}
\value{
An area graph and area chart demonstrating the amount of people displaced from 2008 to 2021
}
\description{
This function produces an area chart visualizing how the amount of displaced individuals over 13 years for a given country (default: USA)
}
\examples{
displaced_areachart ()
displaced_areachart("Nicaragua")
}
