% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaced_functions.R
\name{displaced_histogram}
\alias{displaced_histogram}
\title{Distribution of the Internally Displaced Individuals Over Thirteen Years (2008-2021)}
\usage{
displaced_histogram(nation = "United States")
}
\arguments{
\item{nation}{Name of country (default: United States)}
}
\value{
A histogram demonstrating the amount of people displaced based on the country selected from 2008 to 2021
}
\description{
This function visualizes the distribution of people internally displaced by natural disasters and climate change in a specific country (default: USA)
}
\examples{
displaced_histogram()
displaced_histogram("Dominican Republic")
}
