\name{export}
\alias{export}
\title{Generic Function to Export Output to Files...}
\usage{export(x, filename, ...)
}
\description{Generic Function to Export Output to Files}
\author{Tobias Verbeke}
\arguments{\item{x}{object to export to a file}
\item{filename}{name of the file to which the output should be exported}
\item{...}{further arguments for the method}
}

