\name{Famin81A}
\alias{Famin81A}
\docType{data}
\title{Famin81A Data}
\description{
  Data with demographic indicators by region of the world
}
\usage{data(Famin81A)}
\format{
  A data frame with 18 observations on the following 5 variables.
  \describe{
    \item{\code{Region}}{a factor with 18 levels giving the region}
    \item{\code{Population.growth}}{a numeric vector}
    \item{\code{Infant.Mortality}}{a numeric vector; Infant mortality}
    \item{\code{Dietary.Energy}}{a numeric vector; Dietary Energy}
    \item{\code{Productivity}}{a numeric vector; Productivity}
  }
}
\references{
  Friday, L. and Laskey, R. (1989). The Fragile Environment, 
  The Darwin College Lecture. Cambridge University Press, UK.
}
\keyword{datasets}
