\name{Famin81A}
\alias{Famin81A}
\alias{Famin81A}
\title{Famin81A Data...}
\description{Famin81A Data
Data with demographic indicators by region of the world}
\docType{data}
\references{Friday, L. and Laskey, R. (1989). The Fragile Environment, The
Darwin College Lecture. Cambridge University Press, UK.}
\keyword{datasets}

