\name{print.mpm}
\alias{print.mpm}
\title{Print Method for mpm Objects...}
\usage{print.mpm(x, digits=3, ...)
}
\description{Print Method for mpm Objects}
\seealso{\code{\link{print.default}}}
\value{x is returned invisibly}
\arguments{\item{x}{object of class mpm}
\item{digits}{minimum number of significant digits to be printed}
\item{...}{further arguments for the print method (for printing the contributions)}
}

