\name{print.summary.mpm}
\alias{print.summary.mpm}
\title{Print Method for summary...}
\usage{\method{print}{summary.mpm}(x, digits=2, what=c("columns", "rows", "all"), ...)
}
\description{Print Method for summary.mpm Objects}
\value{x is returned invisibly}
\seealso{\code{\link{print.default}}}
\arguments{\item{x}{object of class summary.mpm}
\item{digits}{minimum number of significant digits to print, defaults to 2}
\item{what}{one of \code{"columns"} (default), \code{"rows"} or \code{"all"}, specifying respectively
whether columns, rows or both need to be printed}
\item{...}{further arguments for the print method}
}

